немного подправил

+++ lab4/main.cpp
@@ -6,7 +6,7 @@ int main()
     vector<int> buf;
     buf.reserve(20);
 
-    for (int i = 0; i < 20; ++i) 
+    for (size_t i = 0; i < 20; ++i) 
     {
         buf.push_back(i); 
     }
@@ -14,10 +14,6 @@ int main()
     Btree<int> tr {buf}; 
     auto tr2 = tr;
 
-
-
     tr.show();
     tr2.show();
-
-    
 }
